// ui/src/generated/BounceTarget.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ClipId } from "./ClipId";
import type { TrackId } from "./TrackId";

/**
 * Where a bounce goes.
 */
export type BounceTarget = { "type": "NewTrack", track: TrackId, clip: ClipId, } | { "type": "InPlace", clip: ClipId, };